backend/requirements.txt
Django==2.2.19
djangorestframework==3.12.4
django-filter==21.1
gunicorn==20.0.4
pytz==2022.4
asgiref==3.2.10
psycopg2-binary==2.9
sqlparse==0.4.3
python-dotenv==0.20.0
Pillow==9.2.0
django-colorfield==0.7.2
drf_extra_fields==3.4.0
djoser==2.1.0
requests==2.26.0
zipp==3.1.0
gunicorn==20.0.4
python-dotenv
drf-base64==2.0